refactor: utils 함수로 텍스트 변환 함수 분리
